refactor: project structure update #1
