+++ src/App.jsx
@@ -1,25 +1,25 @@
-import Home from "./pages/Home";
-import Layout from "./pages/Layout";
-import Details from "./pages/Details";
-import Rankings from "./pages/Rankings";
-import Marketplace from "./pages/Marketplace";
-import CreateAccount from "./pages/CreateAccount";
+import Layout from "./component/common/Layout";
+import Home from "./component/Home";
+import Details from "./component/Details";
+import Rankings from "./component/Rankings";
+import Marketplace from "./component/Marketplace";
+import CreateAccount from "./component/CreateAccount";
 import { createRoutesFromElements } from "react-router";
 import { RouterProvider, createBrowserRouter, Route } from "react-router-dom";
 
 const router = createBrowserRouter(
-	createRoutesFromElements(
-		<Route path="/" element={<Layout />}>
-			<Route index element={<Home />} />
-			<Route path="marketplace" element={<Marketplace />} />
-			<Route path="rankings" element={<Rankings />} />
-			<Route path="create-account" element={<CreateAccount />} />
-			<Route path="details" element={<Details />} />
-		</Route>
-	)
+  createRoutesFromElements(
+    <Route path="/" element={<Layout />}>
+      <Route index element={<Home />} />
+      <Route path="marketplace" element={<Marketplace />} />
+      <Route path="rankings" element={<Rankings />} />
+      <Route path="create-account" element={<CreateAccount />} />
+      <Route path="details" element={<Details />} />
+    </Route>
+  )
 );
 function App() {
-	return <RouterProvider router={router} />;
+  return <RouterProvider router={router} />;
 }
 
 export default App;
